Move useSelector,useDispatch inside the components
