+++ src/hooks/index.js
@@ -1,2 +1,2 @@
 export { useClient } from '@plone/volto/hooks/client/useClient';
-export * from '@plone/volto/hooks/emailNotification/useEmailNotification';
+
